Add script for building a botnet index file
This index file will be used to determine which servers should be controlled. This script will fill the index with all servers we are currently capable of hacking and controlling. List of servers is stored as an array converted into a JSON object written into a text file. While this script automates building the index, this list can be manually edited if necessary.